First commit of the concealment MiTM attacks. Before having a concealment attack, we will re arrange the attacks in the following way: - DHALSIM 1.0 MiTM attack will be now cppo server attack, with the same functionality - naive MiTM attack will be the simple attack that does not filter manipulated packets by tag - MiTM will not have concealmente, but will have an additional parameter that indicate the tag to be manipulated - Concealment MiTM will have the functionality of MiTM and additional parameters to indicate the concealment type and parameters (read from csv, polyline, fixed value, offset of original value, etc)
