+++ dhalsim/python2/automatic_attacker.py
@@ -70,7 +70,7 @@ class AttackerControl(NodeControl):
         """Start a attack process."""
         generic_plc_path = None
         if self.this_attacker_data['type'] == 'mitm':
-            generic_plc_path = Path(__file__).parent.parent.absolute() / "network_attacks" / "mitm_attack.py"
+            generic_plc_path = Path(__file__).parent.parent.absolute() / "network_attacks" / "cppo_server_mitm_attack.py"
         elif self.this_attacker_data['type'] == 'naive_mitm':
             generic_plc_path = Path(__file__).parent.parent.absolute() / "network_attacks" / "naive_attack.py"
         elif self.this_attacker_data['type'] == 'simple_stale':
@@ -78,8 +78,7 @@ class AttackerControl(NodeControl):
         elif self.this_attacker_data['type'] == 'simple_dos':
             generic_plc_path = Path(__file__).parent.parent.absolute() / "network_attacks" / "simple_dos_attack.py"
         elif self.this_attacker_data['type'] == 'concealment_mitm':
-            #generic_plc_path = Path(__file__).parent.parent.absolute() / "network_attacks" / "concealment_mitm.py"
-            raise NoSuchAttack("Attack {attack} does not exists.".format(attack=self.this_attacker_data['type']))
+            generic_plc_path = Path(__file__).parent.parent.absolute() / "network_attacks" / "concealment_mitm.py"
         else:
             raise NoSuchAttack("Attack {attack} does not exists.".format(attack=self.this_attacker_data['type']))
